Correct the OPTICAL_AXIS_OFFSET_X value and move the mounts appropriately.
After the OPTICAL_AXIS_OFFSET_X value is changed, some of the mounts have been repositioned in the printing layout.  The camera mount needs to be moved to the third printing batch, and so the gears are also moved to that batch.

Some of the "wires" connecting the parts that are moved also need to be repositioned.

The pressure angle for the gears is corrected.